move threaded execution to wingman level and rename skill's gpt_call to llm_call to match wording in wingman
